requirements.txt
z3-solver
PyYAML